second commit

+++ src/main/java/com/academic/fh/repository/MovimientoInventarioRepository.java
@@ -3,6 +3,6 @@ package com.academic.fh.repository;
 import com.academic.fh.model.MovimientoInventario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Integer> {
+public interface MovimientoInventarioRepository extends JpaRepository<MovimientoInventario, Long> {
 
 }
