+++ src/main/java/com/academic/fh/repository/ProductoRepository.java
@@ -3,6 +3,6 @@ package com.academic.fh.repository;
 import com.academic.fh.model.Producto;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductoRepository extends JpaRepository<Producto, Integer> {
+public interface ProductoRepository extends JpaRepository<Producto, Long> {
 
 }
